Adding CORS Middleware

+++ routes/event.py
@@ -28,6 +28,7 @@ async def retrieve_event(id: PydanticObjectId) -> Event:
 
 @event_router.post("/new")
 async def create_event(body:Event, user: str = Depends(authenticate)) -> dict:
+    body.creator = user
     await event_database.save(body)
     return {
         "message": "Event created successfully"
@@ -35,6 +36,12 @@ async def create_event(body:Event, user: str = Depends(authenticate)) -> dict:
 
 @event_router.put("/{id}", response_model=Event)
 async def update_event(id: PydanticObjectId, body: EventUpdate = Body(...), user: str = Depends(authenticate)) -> Event:
+    event = await event_database.get(id)
+    if event.creator != user:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Operation not alowed"
+        )
     updated_event = await event_database.update(id, body)
     if not updated_event:
         raise HTTPException(
@@ -45,6 +52,12 @@ async def update_event(id: PydanticObjectId, body: EventUpdate = Body(...), user
 
 @event_router.delete("/{id}")
 async def delete_event(id: PydanticObjectId, user: str = Depends(authenticate)) -> dict:
+    event = await event_database.get(id)
+    if event.creator != user:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Operation not alowed"
+        )
     event = await event_database.delete(id)
     if not event:
         raise HTTPException(
